Turning cursor on/off, wrap cursor around the window when we are rotating, ESC to go back to debug camera mode.
Proper first person and mouse rotation modes with camera hidden.
Only import mouse motion to the GUI.
When turning of the cursor, signal to the GUI to release buttons and
also remove cursor pos.
Reset component on selection manager, so when we turn the cursor off we
can signal to the GUI to un-hover the current selected component.