Add camera clipping functionality, store a fake (buffered) camera position in the camera class
Update the fake position when we make camera position updates, based on
movement and rotation, then before do the camera clipping fetch this
fake position and then work out a non-clipped position that we can use
and pass to the camera class that is used when OpenGL positions the
camera.

Maybe rename the 'fake' position variable in the camera class to make it
more obvious that this is used for pre-clipping state.