Update player alpha based on actual camera position, not zoom amount. Dynamically adjust clipping increment based on distance to player
Always allow the clipping to succeed irrespective of the number of
iterations we decide is the max, since the incremental amount that we
add on each iteration will be based on the max number of iterations.

Dont do camera clipping in debug camera or mode.
Increase the default zoom level at start.